DRY out the backends
Moves the temp dir creation and cleanup to a new download method in
sequins proper.
